Fix zeroing CC in bidirectional mode
