+++ go.mod
@@ -1,6 +1,6 @@
 module github.com/gethiox/HIDI
 
-go 1.16
+go 1.18
 
 require (
 	github.com/d2r2/go-hd44780 v0.0.0-20181002113701-74cc28c83a3e
@@ -9,11 +9,16 @@ require (
 	github.com/fsnotify/fsnotify v1.5.1
 	github.com/go-ini/ini v1.66.4
 	github.com/holoplot/go-evdev v0.0.0-20220421121408-c80ef6a93985
-	github.com/kr/pretty v0.1.0 // indirect
 	github.com/moutend/go-midi v0.3.1
-	github.com/op/go-logging v0.0.0-20160315200505-970db520ece7 // indirect
 	github.com/stretchr/testify v1.7.0
+	gopkg.in/yaml.v3 v3.0.0-20210107192922-496545a6307b
+)
+
+require (
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/kr/pretty v0.1.0 // indirect
+	github.com/op/go-logging v0.0.0-20160315200505-970db520ece7 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
 	golang.org/x/sys v0.0.0-20220412211240-33da011f77ad // indirect
 	gopkg.in/check.v1 v1.0.0-20180628173108-788fd7840127 // indirect
-	gopkg.in/yaml.v3 v3.0.0-20210107192922-496545a6307b
 )
